refactor: 전역적으로 쓰이는 컴포넌트 및 더미데이터 app폴더의 _component로 이동
